requirements.txt
langchain
langchain_openai
langchain_community
wikipedia
pillow
streamlit